improves how session's speakers are printed
